backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.46
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.20
pydantic-settings==2.7.1
python-dotenv==1.0.1
